+++ src/getpodia/course.nim
@@ -171,10 +171,9 @@ func thumbnail*(meta: VideoMeta): VideoMediaMeta {.inline.} =
   meta.getMedia "still_image"
 func hdVideo*(meta: VideoMeta): VideoMediaMeta {.inline.} =
   ## Video HD source
-  meta.getMedia "hd_mp4_video"
-func originalVideo*(meta: VideoMeta): VideoMediaMeta {.inline.} =
-  ## Original video source
-  meta.getMedia "original"
+  result = meta.getMedia "hd_mp4_video"
+  if result.url.len == 0:
+    result = meta.getMedia "md_mp4_video"
 
 when isMainModule:
   from std/tables import pairs